# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Multi-Vendor E-Commerce API",
    "version": "1.0.0",
    "description": "REST API for the multi-vendor e-commerce platform. Most endpoints require a Bearer JWT — click **Authorize** and enter `Bearer <access_token>`."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Current API host"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from POST /auth/login. Enter: `Bearer <access_token>`"
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Secure-access_token",
        "description": "HttpOnly web session cookie issued when X-Auth-Mode is cookie."
      }
    },
    "schemas": {
      "Pagination": {
        "type": "object",
        "required": [
          "total",
          "page",
          "limit",
          "totalPages"
        ],
        "properties": {
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          }
        }
      },
...
      "ConnectOnboardingSuccess": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "required": [
              "url"
            ],
            "properties": {
              "url": {
                "type": "string",
                "format": "uri"
              }
            }
          }
        }
      },
      "ConnectStatusSuccess": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "required": [
              "onboardingStatus",
              "chargesEnabled",
              "payoutsEnabled",
              "detailsSubmitted"
            ],
            "properties": {
              "onboardingStatus": {
                "type": "string",
                "enum": [
                  "NOT_STARTED",
                  "PENDING",
                  "COMPLETE",
                  "RESTRICTED"
                ]
              },
              "chargesEnabled": {
                "type": "boolean"
              },
              "payoutsEnabled": {
                "type": "boolean"
              },
              "detailsSubmitted": {
                "type": "boolean"
              }
            }
          }
        }
      },
      "EarningsAmounts": {
        "type": "object",
        "required": [
          "count",
          "grossAmount",
          "commissionAmount",
          "netAmount"
        ],
        "properties": {
          "count": {
            "type": "integer"
          },
          "grossAmount": {
            "type": "number"
          },
          "commissionAmount": {
            "type": "number"
          },
          "netAmount": {
            "type": "number"
          }
        }
      },
      "EarningsSummarySuccess": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "required": [
              "pending",
              "transferred",
              "failed",
              "reversed"
            ],
            "properties": {
              "pending": {
                "$ref": "#/components/schemas/EarningsAmounts"
              },
              "transferred": {
                "$ref": "#/components/schemas/EarningsAmounts"
              },
              "failed": {
                "$ref": "#/components/schemas/EarningsAmounts"
              },
              "reversed": {
                "$ref": "#/components/schemas/EarningsAmounts"
              }
            }
          }
        }
      },
      "VendorEarning": {
        "type": "object",
        "required": [
          "id",
          "vendorProfileId",
          "vendorOrderId",
          "orderId",
          "grossAmount",
          "commissionRate",
          "commissionAmount",
          "netAmount",
          "currency",
          "status",
          "stripeTransferId",
          "transferredAt",
          "createdAt",
          "updatedAt",
          "order"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "vendorProfileId": {
            "type": "string"
          },
          "vendorOrderId": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "grossAmount": {
            "type": "string"
          },
          "commissionRate": {
            "type": "string"
          },
          "commissionAmount": {
            "type": "string"
          },
          "netAmount": {
            "type": "string"
          },
          "currency": {
            "type": "string",
            "enum": [
              "USD",
              "EUR",
              "GBP",
              "INR",
              "CAD",
              "AUD"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "TRANSFERRED",
              "FAILED",
              "REVERSED"
            ]
          },
          "stripeTransferId": {
            "type": "string",
            "nullable": true
          },
          "transferredAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "order": {
            "type": "object",
            "required": [
              "orderNumber"
            ],
            "properties": {
              "orderNumber": {
                "type": "string"
              }
            }
          }
        }
      },
      "VendorEarningsSuccess": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "required": [
              "earnings",
              "pagination"
            ],
            "properties": {
              "earnings": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/VendorEarning"
                }
              },
              "pagination": {
                "$ref": "#/components/schemas/Pagination"
              }
            }
          }
        }
      },
...
      "VendorPayoutsSuccess": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "required": [
              "payouts",
              "pagination"
            ],
            "properties": {
              "payouts": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/VendorPayout"
                }
              },
              "pagination": {
                "$ref": "#/components/schemas/Pagination"
              }
            }
          }
        }
      },
      "VendorPayoutCommissionSuccess": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "required": [
              "vendorId",
              "commissionRate"
            ],
            "properties": {
              "vendorId": {
                "type": "string"
              },
              "commissionRate": {
                "type": "number"
              }
            }
          }
        }
      },
...
      "ApiError": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ],
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "example": "email"
                },
                "message": {
                  "type": "string",
                  "example": "Invalid email address"
                }
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    },
    {
      "CookieAuth": []
    }
  ],
...
  "paths": {
...
    "/vendor-payouts/webhook": {
      "post": {
        "tags": [
          "Vendor Payouts"
        ],
        "summary": "Stripe Connect webhook (Stripe servers only)",
        "description": "Receives Stripe Connect webhook events (e.g. `account.updated`, `payout.paid`). **Do not call this endpoint directly** — it is for Stripe's servers only. This route is mounted before the global rate limiter so Stripe retries are never throttled.\n",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Stripe Event object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event received"
          },
          "400": {
            "description": "Invalid Stripe signature"
          }
        }
      }
    },
    "/vendor-payouts/connect/onboard": {
      "post": {
        "tags": [
          "Vendor Payouts"
        ],
        "summary": "Start Stripe Connect onboarding (approved Vendors only)",
        "description": "Generates a single-use Stripe-hosted onboarding URL using backend-configured return and refresh URLs; clients cannot supply either redirect. Navigate the browser at top level, then reconcile completion with GET /vendor-payouts/connect/status after Stripe returns.",
        "responses": {
          "200": {
            "description": "Onboarding URL",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "message": "Onboarding link generated",
                  "data": {
                    "url": "https://connect.stripe.com/setup/s/..."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ConnectOnboardingSuccess"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — must be an approved vendor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Vendor profile not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/vendor-payouts/connect/onboard/refresh": {
      "get": {
        "tags": [
          "Vendor Payouts"
        ],
        "summary": "Refresh the Stripe Connect onboarding link (approved Vendors only)",
        "description": "Returns a fresh single-use onboarding URL when Stripe reaches the backend-configured refresh route. Call once per refresh event, navigate at top level, and guard the client against retry loops. Redirect targets are server configuration, never request input.",
        "responses": {
          "200": {
            "description": "Refreshed onboarding URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectOnboardingSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Stripe onboarding has not been started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — must be an approved vendor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Vendor profile not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/vendor-payouts/connect/status": {
      "get": {
        "tags": [
          "Vendor Payouts"
        ],
        "summary": "Get Stripe Connect account status (Vendor only)",
        "description": "Reconciles the Stripe account after onboarding return. Use onboardingStatus, chargesEnabled, payoutsEnabled, and detailsSubmitted as authoritative UI inputs; return query parameters are not proof of completion.",
        "responses": {
          "200": {
            "description": "Connect account status",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "message": "Connect status fetched",
                  "data": {
                    "onboardingStatus": "COMPLETE",
                    "chargesEnabled": true,
                    "payoutsEnabled": true,
                    "detailsSubmitted": true
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ConnectStatusSuccess"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — VENDOR role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/vendor-payouts/earnings": {
      "get": {
        "tags": [
          "Vendor Payouts"
        ],
        "summary": "List vendor earnings (approved Vendors only)",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10,
              "maximum": 100
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "TRANSFERRED",
                "FAILED",
                "REVERSED"
              ]
            }
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated earnings list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorEarningsSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination, status, or date filter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/vendor-payouts/earnings/summary": {
      "get": {
        "tags": [
          "Vendor Payouts"
        ],
        "summary": "Get earnings summary (approved Vendors only)",
        "description": "Returns count, gross, commission, and net aggregates for every earning status.",
        "responses": {
          "200": {
            "description": "Earnings summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EarningsSummarySuccess"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — must be an approved vendor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Vendor profile not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/vendor-payouts/payouts": {
      "get": {
        "tags": [
          "Vendor Payouts"
        ],
        "summary": "List payouts (approved Vendors only)",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10,
              "maximum": 100
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "PAID",
                "FAILED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated payouts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorPayoutsSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination or status filter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/vendor-payouts/admin/commission/{vendorId}": {
      "patch": {
        "tags": [
          "Vendor Payouts"
        ],
        "summary": "Set a vendor's commission rate (Admin only)",
        "parameters": [
          {
            "in": "path",
            "name": "vendorId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Vendor user ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "commissionRate"
                ],
                "properties": {
                  "commissionRate": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 100,
                    "description": "Platform commission percentage (0-100)",
                    "example": 15
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Commission rate updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorPayoutCommissionSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — ADMIN role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Vendor not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  }
}